fix: 페이지 위치 변경, ui 리팩토링
